+++ docs/source/changelog.rst
@@ -1,5 +1,11 @@
 Change Log
 ==========
+
+Unreleased Version 0.9.3 - TBD
+------------------------------
+    * UUID objects can now be serialized
+      (`#130 <https://github.com/jsonpickle/jsonpickle/issues/130>`_).
+
 Version 0.9.2 - March 20, 2015
 ------------------------------
     * Fixes for serializing objects with custom handlers.